+++ backend/app/models.py
@@ -1,0 +1,42 @@
+"""
+Pydantic models for request/response validation
+"""
+from pydantic import BaseModel
+from typing import Optional, List
+
+
+class Ingredient(BaseModel):
+    quantity: float
+    unit: str
+    name: str
+
+
+class RecipeCreate(BaseModel):
+    name: str
+    category: str
+    main_ingredients: List[Ingredient]
+    common_ingredients: List[str]
+    instructions: str
+    prep_time: int
+    portions: int
+
+
+class RecipeUpdate(BaseModel):
+    name: Optional[str] = None
+    category: Optional[str] = None
+    main_ingredients: Optional[List[Ingredient]] = None
+    common_ingredients: Optional[List[str]] = None
+    instructions: Optional[str] = None
+    prep_time: Optional[int] = None
+    portions: Optional[int] = None
+
+
+class RecipeResponse(BaseModel):
+    id: int
+    name: str
+    category: str
+    main_ingredients: List[Ingredient]
+    common_ingredients: List[str]
+    instructions: str
+    prep_time: int
+    portions: int
